Move out ErrorModal to components
